separated styles and data from main component file
