+++ src/containers/MyComplaints/dummyData/index.js
@@ -1,0 +1,55 @@
+let DATA = {
+  category: 'seguridad',
+  title: 'Abuso Animal',
+  description:
+    'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
+  date: 'Ocurrido el 6 mayo, 2021',
+  location:
+    'Hacienda Los Fresnos, Av. Los Fresnos #345, Jalisco, Tlatomulco de Zúñiga',
+  author: 'Denuncia por José Támara',
+  authorDate: '7 de mayo, 2021',
+  fav: false,
+};
+
+let MyCompEmpty = [];
+
+let MyComp = [
+  {
+    id: 1,
+    category: 'seguridad',
+    title: 'Abuso animal',
+    date: '7 de mayo, 2021',
+  },
+  {
+    id: 2,
+    category: 'Medio ambiente',
+    title: 'Tala clandestina de árboles',
+    date: '3 de mayo, 2021',
+  },
+  {
+    id: 3,
+    category: 'Transporte',
+    title: 'Abuso animal',
+    date: '1 de mayo, 2021',
+  },
+  {
+    id: 4,
+    category: 'Transporte',
+    title: 'Abuso animal',
+    date: '1 de mayo, 2021',
+  },
+  {
+    id: 5,
+    category: 'Transporte',
+    title: 'Abuso animal',
+    date: '1 de mayo, 2021',
+  },
+  {
+    id: 6,
+    category: 'Transporte',
+    title: 'Abuso animal',
+    date: '1 de mayo, 2021',
+  },
+];
+
+export {DATA, MyComp, MyCompEmpty};
